Refactor code (no current changes to functionality)
