+++ src/util.ts
@@ -1,5 +1,6 @@
-import * as fs from 'fs-extra'
 import * as path from 'path'
+
+import * as fs from 'fs-extra'
 
 export async function touch(p: string) {
   try {
